+++ app/login.php
@@ -7,17 +7,17 @@ use Illuminate\Database\Eloquent\Model;
 class login extends Model
 
 {
-    public $timestamps = false;
-
-    protected $fillable= array('id','name','from', 'password');
-
-    public function getData()
-    {
-        $id = $this -> id;
-        $name = $this -> name;
-        $from = $this -> from;
-        $password = $this -> password;
-
-        return compact('id','name', 'from', 'password');
-    }
+//    public $timestamps = false;
+//
+//    protected $fillable= array('id','name','from', 'password');
+//
+//    public function getData()
+//    {
+//        $id = $this -> id;
+//        $name = $this -> name;
+//        $from = $this -> from;
+//        $password = $this -> password;
+//
+//        return compact('id','name', 'from', 'password');
+//    }
 }
