+++ resources/views/welcome.blade.php
@@ -88,7 +88,6 @@
                     <a href="https://medrt.co.jp/">Homepage</a>
                     <a href="/contact">contact</a>
                     <a href="/about">about</a>
-                    <a href="/y_list">y_list</a>
                     <a href="https://github.com/laravel/laravel">公式GitHub</a>
                     <a href="/new">new</a>
                     <a href="/search">search</a>
